+++ ios/Classes/handlers/Initialization/StartManually.swift
@@ -14,6 +14,10 @@ final class StartManuallyRequest: Request {
         if let baseUrl = arguments.baseUrl {
             ApphudHttpClient.shared.domainUrlString = baseUrl
         }
+        if let user = Apphud.currentUser() {
+            result(user.toMap())
+            return
+        }
         Apphud.startManually(apiKey: arguments.apiKey,
                                  userID: arguments.userID,
                                  deviceID: arguments.deviceID,